# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "efgs-federation-gateway",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0-SNAPSHOT"
  },
  "servers": [{
      "url": "http://localhost:8090",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/diagnosiskeys/download/{date}": {
      "get": {
        "tags": ["Diagnosis Keys Exchange Interface", "Download"],
        "summary": "Downloads diagnosis keys dataset by date.",
        "description": "Downloads the latest data by date. The date indicates the start point for query. Means the last 5 Days are currentDate-5",
        "operationId": "downloadDiagnosisKeys",
        "parameters": [{
            "name": "date",
            "in": "path",
            "description": "Date from where the query should start until today.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2020-07-31"
          }, {
            "name": "batchTag",
            "in": "header",
            "description": "Optional Tag to submit the last received batchTag of the day.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "20200731-1"
          },
          {
            "name": "batchIndex",
            "in": "header",
            "description": "Optional Index. Downloads just a part of the batch together with the signature.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "1,2,3,4 etc."
          }
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "headers": {
              "batchTag": {
                "description": "Tag of the batch.",
                "required": true,
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              },
              "nextBatchTag": {
                "description": "Tag of the next available batch of the day. Has the value \"null\" if no further BatchTag exists for requested date",
                "required": true,
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              },
              "nextBatchIndex": {
                "description": "Next available Index. If null, next batchtag has to be used.",
                "required": false,
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              },
              "batchSignature": {
                "description": "Batch Signature. Only available with index.",
                "required": false,
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json+v1.0": {
                "example": "diagnosisKeyBatch"
              }
            }
          },
          "400": {
            "description": "Invalid or missing request header."
          },
          "403": {
            "description": "Forbidden call in cause of missing or invalid client certificate."
          },
          "404": {
            "description": "BatchTag not found or no data exists."
          },
          "406": {
            "description": "Data format or content is not valid."
          },
          "410": {
            "description": "Date for download expired. Date does not more exists."
          }
        }
      }
    },
...
  },
...
}